+++ Sources/TOMLKit/Encoder/SingleValueEncodingContainer.swift
@@ -37,155 +37,119 @@ extension InternalTOMLEncoder.SVEC {
 	func encode(_ value: Float) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = Double(value)
+				t.table?[self.parentKey!.stringValue] = Double(value).tomlValue
 			case let .right((array, index)):
-				array[index] = Double(value)
+				array[index] = Double(value).tomlValue
 		}
 	}
 
 	func encode(_ value: Int) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = value
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = value
 		}
 	}
 
 	func encode(_ value: Int8) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: Int16) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: Int32) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: Int64) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: UInt) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: UInt8) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: UInt16) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: UInt32) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode(_ value: UInt64) throws {
 		switch self.tomlValueOrArray {
 			case let .left(t):
-				t.table?[self.parentKey!.stringValue] = value.tomlInt
+				t.table?[self.parentKey!.stringValue] = Int(value)
 			case let .right((array, index)):
-				array[index] = value.tomlInt
+				array[index] = Int(value)
 		}
 	}
 
 	func encode<T>(_ value: T) throws where T: Encodable {
-		switch value {
-			case is TOMLTime:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLTime
-					case let .right((array, index)):
-						array[index] = value as! TOMLTime
-				}
-			case is TOMLDate:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLDate
-					case let .right((array, index)):
-						array[index] = value as! TOMLDate
-				}
-			case is TOMLDateTime:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLDateTime
-					case let .right((array, index)):
-						array[index] = value as! TOMLDateTime
-				}
-			case is TOMLTable:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLTable
-					case let .right((array, index)):
-						array[index] = value as! TOMLTable
-				}
-			case is TOMLArray:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLArray
-					case let .right((array, index)):
-						array[index] = value as! TOMLArray
-				}
-			case is TOMLInt:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						t.table?[self.parentKey!.stringValue] = value as! TOMLInt
-					case let .right((array, index)):
-						array[index] = value as! TOMLInt
-				}
-			default:
-				switch self.tomlValueOrArray {
-					case let .left(t):
-						let encoder = InternalTOMLEncoder(.left(t.tomlValue), parentKey: self.parentKey, codingPath: self.codingPath, userInfo: self.userInfo)
-						try value.encode(to: encoder)
-					case let .right((array, index)):
-						let encoder = InternalTOMLEncoder(.right((array: array, index: index)), parentKey: self.parentKey, codingPath: self.codingPath, userInfo: self.userInfo)
-						try value.encode(to: encoder)
-				}
+		if value is TOMLValueConvertible {
+			switch self.tomlValueOrArray {
+				case let .left(t):
+					t.table?[self.parentKey!.stringValue] = (value as! TOMLValueConvertible)
+				case let .right((array, index)):
+					array[index] = (value as! TOMLValueConvertible)
+			}
+		} else {
+			switch self.tomlValueOrArray {
+				case let .left(t):
+					let encoder = InternalTOMLEncoder(.left(t.tomlValue), parentKey: self.parentKey, codingPath: self.codingPath, userInfo: self.userInfo)
+					try value.encode(to: encoder)
+				case let .right((array, index)):
+					let encoder = InternalTOMLEncoder(.right((array: array, index: index)), parentKey: self.parentKey, codingPath: self.codingPath, userInfo: self.userInfo)
+					try value.encode(to: encoder)
+			}
 		}
 	}
 }